requirements.txt
boto3==1.17.112
Pillow==8.3.2